on going misaligned memory problem fixing

+++ Common.h
@@ -3,21 +3,22 @@
 #include<cmath>
 #include<memory>
 
+#include<cuda.h>
 #include<cuda_runtime.h>
 #include<thrust/device_vector.h>
 using std::shared_ptr;
 using std::make_shared;
 
-__constant__ const double INF = std::numeric_limits<double>::infinity();
-__constant__ const double PI = 3.1415926535897932385;
+__constant__ const float INF = std::numeric_limits<float>::infinity();
+__constant__ const float PI = 3.1415926535897932385;
 
-inline __device__ double Deg2Rad(double degrees)
+inline __device__ float Deg2Rad(float degrees)
 {
 	return degrees * PI / 180.0;
 }
 
 
-inline double Clamp(double x, double min, double max)
+inline float Clamp(float x, float min, float max)
 {
 	if (x < min)
 	{
@@ -33,4 +34,4 @@ inline double Clamp(double x, double min, double max)
 }
 
 #include"Vec3.h"
-#include"Ray.h"
+#include"Ray.cuh"
